smart_attendance_python/requirements.txt
fastapi
uvicorn
numpy
Pillow
opencv-python
mysql-connector-python
python-multipart
python-dotenv
